Clarify timing module boundaries
