+++ modal_training_gym/common/step_timing.py
@@ -1,121 +1,6 @@
-from __future__ import annotations
+from modal_training_gym.common.sync_timing_recording import record_sync_timing_event
+from modal_training_gym.common.timing_types import Substep
 
-from collections.abc import MutableMapping
-from enum import Enum
-from typing import Literal, NotRequired, TypeAlias, TypedDict
+record_step_time_event = record_sync_timing_event
 
-from modal_training_gym.common.status import SlimeStatus
-
-TimingLane: TypeAlias = Literal[
-    "rollout",
-    "reward",
-    "training",
-    "coordination",
-]
-
-
-class StepTiming(TypedDict):
-    start: int | None
-    end: int | None
-    duration_s: int | None
-
-
-class SubstepTimingInterval(TypedDict):
-    start: float
-    duration_s: float
-    step_id: NotRequired[int]
-    training_role: NotRequired[str]
-    training_rank: NotRequired[int]
-    slowest_rank: NotRequired[int]
-    training_world_size: NotRequired[int]
-    reported_rank_count: NotRequired[int]
-    active_duration_s: NotRequired[float]
-    timeline_lane: NotRequired[TimingLane]
-    parent_phase: NotRequired[str]
-    display_name: NotRequired[str]
-
-
-class SubstepTiming(TypedDict):
-    start: float
-    duration_s: float | None
-    intervals: NotRequired[list[SubstepTimingInterval]]
-
-
-StepTimes: TypeAlias = dict[str, StepTiming]
-SubstepTimes: TypeAlias = dict[str, dict[str, SubstepTiming]]
-
-
-class Substep(str, Enum):
-    EVAL_BEFORE = SlimeStatus.EVAL_ROLLOUT_LOGGING.value
-    GENERATE_ROLLOUTS = SlimeStatus.ROLLOUT_LOGGING.value
-    OFFLOAD_ROLLOUT = SlimeStatus.OFFLOAD_ROLLOUT.value
-    COMPUTE_LOG_PROBS = SlimeStatus.COMPUTE_LOG_PROBS.value
-    OPTIMIZER_STEP = SlimeStatus.OPTIMIZER_STEP.value
-    CHECKPOINT_SAVE = SlimeStatus.CHECKPOINT_SAVE.value
-    OFFLOAD_TRAIN = SlimeStatus.OFFLOAD_TRAIN.value
-    WEIGHT_SYNC = SlimeStatus.WEIGHT_SYNC.value
-    EVAL_AFTER = f"{SlimeStatus.EVAL_ROLLOUT_LOGGING.value}_end"
-
-
-class TrainingSubstep(str, Enum):
-    POLICY_LOG_PROBS = SlimeStatus.COMPUTE_LOG_PROBS.value
-    REFERENCE_LOG_PROBS = "reference_log_probs"
-    TEACHER_LOG_PROBS = "teacher_log_probs"
-    FORWARD_BACKWARD = "forward_backward"
-    OPTIMIZER_STEP = SlimeStatus.OPTIMIZER_STEP.value
-
-
-def record_step_time_event(
-    step_times: MutableMapping[str, float],
-    training_run_id: str,
-    current_step: object,
-    phase: str,
-    step_event: str,
-    event_ts: float,
-) -> None:
-    if not (isinstance(current_step, int) and current_step > 0):
-        return
-    event_time = round(float(event_ts), 3)
-    step_window_start_key = f"{training_run_id}:{current_step}:substep_start"
-
-    def record_first_in_step_window(
-        key: str, *, allow_before_step_window: bool = False
-    ) -> None:
-        existing = step_times.get(key)
-        raw_step_window_start = step_times.get(step_window_start_key)
-        step_window_start = (
-            float(raw_step_window_start) if raw_step_window_start is not None else None
-        )
-        if existing is not None and (
-            step_window_start is None or float(existing) >= step_window_start
-        ):
-            return
-
-        recorded_time = event_time
-        if step_window_start is not None and not allow_before_step_window:
-            recorded_time = max(recorded_time, step_window_start)
-        step_times[key] = recorded_time
-
-    if step_event == "start":
-        step_times[f"{training_run_id}:{current_step}:start"] = event_time
-    elif step_event == "finish":
-        step_times[f"{training_run_id}:{current_step}:finish"] = event_time
-
-    if step_event == "substep_start":
-        step_times[step_window_start_key] = event_time
-    elif step_event == "substep_finish":
-        record_first_in_step_window(f"{training_run_id}:{current_step}:substep_finish")
-    elif step_event in ("eval_begin", "eval_end"):
-        substep = (
-            Substep.EVAL_BEFORE if step_event == "eval_begin" else Substep.EVAL_AFTER
-        )
-        record_first_in_step_window(
-            f"{training_run_id}:{current_step}:substep:{substep.value}",
-            allow_before_step_window=step_event == "eval_begin",
-        )
-    elif not step_event and phase == Substep.EVAL_BEFORE.value:
-        pass
-    elif step_event == "finish":
-        pass
-    else:
-        record_first_in_step_window(f"{training_run_id}:{current_step}:substep:{phase}")
+__all__ = ["Substep", "record_step_time_event"]
