+++ modal_training_gym/frameworks/slime/launcher.py
@@ -21,7 +21,7 @@ import shlex
 import subprocess
 import tempfile
 import time
-from collections.abc import Callable, Hashable, Mapping
+from collections.abc import Callable, Hashable
 from enum import Enum
 from pathlib import Path, PurePosixPath
 from typing import Any, cast
@@ -30,13 +30,13 @@ from modal import App, Dict as ModalDict, Image, Secret, Volume, Retries
 
 from modal_training_gym.common import hf_secrets, proxy_auth_secrets
 
-from modal_training_gym.common.async_step_timing import (
+from modal_training_gym.common.async_timing_aggregation import (
     StepTimingSnapshot,
     aggregate_async_step_times,
     apply_step_timing_snapshot,
     reconcile_attempt_step_times,
 )
-from modal_training_gym.common.async_timing import AsyncTimingEvent
+from modal_training_gym.common.async_timing_types import AsyncTimingEvent
 from modal_training_gym.common.dataset import DatasetConfig, HarborDataset
 from modal_training_gym.common.framework import (
     mount_tools_dir,
@@ -55,7 +55,7 @@ from modal_training_gym.common.run import (
     record_resume_checkpoint,
     torch_dist_resume_checkpoint,
 )
-from modal_training_gym.common.step_timing import StepTimes, SubstepTimes
+from modal_training_gym.common.sync_timing_aggregation import aggregate_sync_step_times
 from modal_training_gym.common.launcher_helpers import (
     build_app_tags,
     build_terminal_run_record,
@@ -72,7 +72,7 @@ from modal_training_gym.common.launcher_helpers import (
 )
 from modal_training_gym.common.wandb import WandbConfig
 from modal_training_gym.common.status import SlimeStatus
-from modal_training_gym.common.step_timing import Substep
+from modal_training_gym.common.timing_types import StepTimes, Substep, SubstepTimes
 from modal_training_gym.utils.metadata import STEP_TIMING_SNAPSHOTS_DICT_NAME
 
 from modal_training_gym.train_recipes.slime_recipe.recipe import (
@@ -290,116 +290,6 @@ def _preflight_wandb(wandb_cfg: WandbConfig) -> str:
     from modal_training_gym.common.wandb import preflight_wandb
 
     return preflight_wandb(wandb_cfg)
-
-
-def aggregate_step_times(
-    step_times_dict: Mapping[str, float | None],
-    run_id: str,
-    num_steps: int,
-    SUBSTEP_ORDER: list[str],
-    OPTIONAL_SUBSTEPS: set[str],
-) -> tuple[StepTimes, SubstepTimes]:
-    """Organize step and substep timings from the Modal dict at the end of a run.
-
-    Records each step's start/end/duration; missing timestamps become None.
-    Each substep's duration is the gap from its start to the next recorded
-    substep's start (or the step's end for the last one). Duration is None
-    if a mandatory substep in between is missing. Step duration is computed
-    independently, since substeps include work outside the step (evals,
-    checkpointing).
-    """
-    step_times: StepTimes = {}
-    substep_times: SubstepTimes = {}
-
-    for current_step_num in range(1, num_steps + 1):
-        start_key = f"{run_id}:{current_step_num}:start"
-        finish_key = f"{run_id}:{current_step_num}:finish"
-
-        raw_start_time = step_times_dict.get(start_key)
-        raw_end_time = step_times_dict.get(finish_key)
-        precise_start_time = (
-            float(raw_start_time) if raw_start_time is not None else None
-        )
-        precise_end_time = float(raw_end_time) if raw_end_time is not None else None
-        start_time = int(precise_start_time) if precise_start_time is not None else None
-        end_time = int(precise_end_time) if precise_end_time is not None else None
-
-        duration = None
-        if start_time is not None and end_time is not None:
-            duration = end_time - start_time
-
-        step_times[f"{current_step_num}"] = {
-            "start": start_time,
-            "end": end_time,
-            "duration_s": duration,
-        }
-
-        raw_step_window_start = step_times_dict.get(
-            f"{run_id}:{current_step_num}:substep_start"
-        )
-        step_window_start = (
-            float(raw_step_window_start) if raw_step_window_start is not None else None
-        )
-        full_step_start_time = (
-            step_window_start if step_window_start is not None else precise_start_time
-        )
-        full_step_end_time = step_times_dict.get(
-            f"{run_id}:{current_step_num}:substep_finish"
-        )
-        if full_step_end_time is not None:
-            full_step_end_time = float(full_step_end_time)
-            if step_window_start is not None and full_step_end_time < step_window_start:
-                full_step_end_time = precise_end_time
-        else:
-            full_step_end_time = precise_end_time
-
-        step_key = str(current_step_num)
-        substep_times[step_key] = {}
-        eval_before = Substep.EVAL_BEFORE.value
-        present: set[str] = set()
-        recorded: list[tuple[float, int, str]] = []
-        for order_idx, substep in enumerate(SUBSTEP_ORDER):
-            substep_start = step_times_dict.get(
-                f"{run_id}:{current_step_num}:substep:{substep}"
-            )
-            if substep_start is None:
-                continue
-            substep_start = float(substep_start)
-            if (
-                step_window_start is not None
-                and substep_start < step_window_start
-                and substep != eval_before
-            ):
-                continue
-            if full_step_start_time is not None and substep != eval_before:
-                substep_start = max(substep_start, full_step_start_time)
-            if full_step_end_time is not None:
-                substep_start = min(substep_start, full_step_end_time)
-            present.add(substep)
-            recorded.append((substep_start, order_idx, substep))
-        recorded.sort()
-
-        for idx, (substep_start, order_idx, substep) in enumerate(recorded):
-            if idx + 1 < len(recorded):
-                next_start, next_idx = recorded[idx + 1][0], recorded[idx + 1][1]
-            else:
-                next_start, next_idx = full_step_end_time, len(SUBSTEP_ORDER)
-
-            gap = SUBSTEP_ORDER[order_idx + 1 : next_idx]
-            dropped_mandatory = any(
-                s not in OPTIONAL_SUBSTEPS and s not in present for s in gap
-            )
-            if next_start is None or dropped_mandatory:
-                substep_duration = None
-            else:
-                substep_duration = round(max(next_start - substep_start, 0.0), 3)
-
-            substep_times[step_key][substep] = {
-                "start": round(substep_start, 3),
-                "duration_s": substep_duration,
-            }
-
-    return step_times, substep_times
 
 
 def build_slime_app(
@@ -1051,7 +941,7 @@ def build_slime_app(
                 for key, value in stored_values.items()
                 if isinstance(key, str) and isinstance(value, (int, float))
             }
-            step_times, substep_times = aggregate_step_times(
+            step_times, substep_times = aggregate_sync_step_times(
                 recorded_times,
                 run_id,
                 num_steps,
