+++ tests/testdata/train_async.py.output
@@ -5,7 +5,7 @@ if '/root' not in _tg_sys.path:
 try:
     from modal_training_gym.frameworks.slime.phase_reporting import (
         flush_async_timing_events as _tg_flush_timings,
-        flush_async_timing_events_on_error as _tg_flush_timings_on_error,
+        flush_async_timing_queue_before_reraise as _tg_flush_timings_on_error,
         report_async_timing_event as _tg_report,
     )
 except ImportError:
